api/requirements.txt
Djang=='1.7'
